requirements.txt
Flask==3.1.0
python-dotenv==1.0.1 
Flask-RESTful==0.3.10 
Flask-SQLAlchemy==3.1.1 
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Testing==0.8.1 
pytest==8.3.5
pytest-flask==1.3.0
requests==2.32.3
python-dateutil
pandas
xlsxwriter